requirements.txt
icrawler==0.6.9
minigpt4==0.0.1.dev0
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
torch==2.5.1
transformers==4.47.1
